refactor: move BOSS code to its own isolated component
