+++ src/components/BOSSTimetable/util/generateICal.ts
