+++ src/components/BOSSTimetable/util/parseMeetings.ts
